Fix many FontStash problems
Includes a big refactor of the way colors are handled (basiclly pushed
all calls to glColor down to as close to the OpenGL stuff as possible,
which makes things a bit clearer).  Also migrated a lot of the text
rendering to the "fixed" functions, which gets us one step closer to
getting rid of the "unfixed" versions.  Fixed a couple of other problems
as well.