Hide initial polywall
Hide polywall that is initially added to editor when dragged from dock.
This is really just an aesthetic thing, and is kind of ugly because the
way walls are rendered and the whole shadow-wall hack to hide the fact
that wall outlines are merged and drawn as a single entity, rather than
as a series of independent objects the way, say, loadout zones are.