Traffic in absolute paths
The motiviation for this change was to make paths shown to the user be
absolute rather than relative to reduce confusion about where the
folders actually are.  This is mostly to help me when I am running
several versions of the game during development and testing, but I think
it will help causal players as well.