Make sure healthbars are always visible, even with dark team colors
Fixes #554
It turns out that neither HSV nor HSL is very good for this purpose.  Instead, I implemented an HSP converter, which does work well (P = Perceived brightness).  We can change the minimum brightness in the Color class.  

Should we also implement this for other objects, like loadout zone markings?  If so, reopen this case, or create a new one.

You can easily see the results of this on the color picker (on the editor>teams menu), where you can change colors and see the effects on a rendered ship and turret.